Fixed error with example on terraform 1.15
Faced the following error when running the example on terraform 1.15 (did not try other versions), also the resource `restapi_object` is missing the alias field for targeting  a specific provider

```
Error: Unsupported block type

  on main.tf line 21, in provider "restapi":
  21:   headers {

Blocks of type "headers" are not expected here. Did you mean to define
argument "headers"? If so, use the equals sign to assign it a value.
```